Split QuickOpen.ts apart
