+++ browser/src/Services/Menu/Menu.less
@@ -81,4 +81,8 @@
         }
     }
 
+    .footer {
+        text-align: center;
+    }
+
 }
