+++ browser/src/Services/QuickOpen/QuickOpenItem.ts
@@ -1,0 +1,61 @@
+// We use basename/dirname for label/detail
+// So let's say you want the label YO with detail a greeting
+// you would make file = "a greeting/YO"
+enum QuickOpenType {
+    bookmark,
+    bookmarkHelp,
+    file,
+    folder,
+    folderHelp,
+    bufferLine,
+    color,
+}
+
+// Wrapper around quick open items, this not only allows us to show multiple icons
+// It also allows us to distinguish what happens once we know their icon
+class QuickOpenItem {
+    // We take a type, and then give an fa icon
+    public static convertTypeToIcon(type: QuickOpenType): string {
+        switch (type) {
+            case QuickOpenType.bookmark:
+                return "star-o"
+            case QuickOpenType.bookmarkHelp:
+                return "info"
+            case QuickOpenType.file:
+                return "file-text-o"
+            case QuickOpenType.folder:
+                return "folder-o"
+            case QuickOpenType.folderHelp:
+                return "folder-open-o"
+            case QuickOpenType.bufferLine:
+                return "angle-right"
+            case QuickOpenType.color:
+                return "paint-brush"
+            default:
+                return "question-circle-o"
+        }
+    }
+
+    // Each has an item, and an icon
+    private _item: string
+    private _icon: string
+    private _lineNu: number
+
+    public get item(): string {
+        return this._item
+    }
+
+    public get icon(): string {
+        return this._icon
+    }
+
+    public get lineNu(): number {
+        return this._lineNu
+    }
+
+    constructor(item: string, type: QuickOpenType, num?: number) {
+        this._item = item
+        this._icon = QuickOpenItem.convertTypeToIcon(type)
+        this._lineNu = num
+    }
+}
